Protobuf files must be kept in correct folder
We must align the Protobuf package ID and the folder in which `.proto` file(s) are found. Indeed, the `excludePaths` on `substreams.yaml` is based on the paths in which they were originally defined. So before this PR, `excludePaths: [instructions.proto]` has to be used, now `excludePaths: ["sol/instructions"] can be used like expected.